Redirect HTTP to HTTPS on the same port
This configuration allows to redirect any HTTP connection to HTTPS, on the same port. This is useful when using a custom port.

I'm adding this config here because it needs to be included in every server configuration, not just the main one, and this file is included in each website configuration.